bug fix horiz multi choice CSS add warning linked text file upload failure
git-svn-id: http://imathas.googlecode.com/svn/trunk@369 c89b4f0b-ac2a-0410-9773-c9071ee4f95d

+++ course/deletelinkedtext.php
@@ -47,7 +47,9 @@ if (!(isset($teacherid))) { // loaded by a NON-teacher
 			if (mysql_num_rows($result)==1) { 
 				$uploaddir = rtrim(dirname(__FILE__), '/\\') .'/files/';
 				$filename = substr($text,5);
-				unlink($uploaddir . $filename);
+				if (file_exists($uploaddir . $filename)) {
+					unlink($uploaddir . $filename);
+				}
 			}
 		}
